+++ modRlp/src/org/aion/rlp/DecodeResult.java
@@ -35,7 +35,7 @@
 package org.aion.rlp;
 
 import java.io.Serializable;
-import org.aion.util.hex.Hex;
+import org.aion.util.conversions.Hex;
 
 @SuppressWarnings("serial")
 public class DecodeResult implements Serializable {
