+++ modRlp/test/org/aion/rlp/RLPDump.java
@@ -35,7 +35,7 @@
 package org.aion.rlp;
 
 import java.util.Arrays;
-import org.aion.util.hex.Hex;
+import org.aion.util.conversions.Hex;
 import org.junit.Test;
 
 public class RLPDump {
