+++ modRlp/test/org/aion/rlp/RLPSpecTest.java
@@ -32,7 +32,7 @@ import static org.aion.util.bytes.ByteUtil.byteArrayToLong;
 import java.math.BigInteger;
 import java.util.List;
 import org.aion.util.bytes.ByteUtil;
-import org.aion.util.hex.Hex;
+import org.aion.util.conversions.Hex;
 import org.junit.Test;
 
 /**
