+++ modUtil/src/module-info.java
@@ -1,5 +1,5 @@
 module aion.util {
     exports org.aion.util;
     exports org.aion.util.bytes;
-    exports org.aion.util.hex;
+    exports org.aion.util.conversions;
 }
